Basic Character
Created the "Player" scene to control shared player values, and created a child scene "Explorer" to have a visible Character that can move.

"Explorer" script currently inherits "Player". This may be changed to instead have the Explorer be a sub-node to the Player so that swapping between Characters does not trigger the relevant scene events.